+++ src/krow/kröw/program/api/callables/ParameterizedTask.java
@@ -1,4 +1,4 @@
-package kröw.app.api.callables;
+package kröw.program.api.callables;
 
 public interface ParameterizedTask<P> {
 	void execute(P param);
