+++ src/krow/kröw/program/api/callables/VarArgsTask.java
@@ -1,4 +1,4 @@
-package kröw.app.api.callables;
+package kröw.program.api.callables;
 
 public interface VarArgsTask<PT> {
 	void execute(@SuppressWarnings("unchecked") PT... params);
