+++ src/krow/kröw/program/api/collections/UniqueArrayList.java
@@ -1,4 +1,4 @@
-package kröw.app.api.collections;
+package kröw.program.api.collections;
 
 import java.util.ArrayList;
 import java.util.Collection;
