+++ src/krow/kröw/program/api/connections/Client.java
@@ -1,4 +1,4 @@
-package kröw.app.api.connections;
+package kröw.program.api.connections;
 
 import java.io.EOFException;
 import java.io.IOException;
