+++ src/krow/kröw/program/api/connections/ClientListener.java
@@ -1,4 +1,4 @@
-package kröw.app.api.connections;
+package kröw.program.api.connections;
 
 public interface ClientListener {
