+++ src/krow/kröw/program/api/connections/FullClientListener.java
@@ -1,4 +1,4 @@
-package kröw.app.api.connections;
+package kröw.program.api.connections;
 
 public interface FullClientListener extends ClientListener {
